# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NutriPlanner API
  version: 1.0.0
  description: Coursework API for nutrition data management, recipe CRUD, and analytics/recommendations.
servers:
  - url: http://localhost:3000/api/v1
tags:
  - name: Auth
  - name: Ingredients
  - name: Recipes
  - name: Analytics
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user and issue tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and issue tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and return new token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
        "401":
          $ref: "#/components/responses/Unauthorized"
  /ingredients:
    get:
      tags: [Ingredients]
      summary: List ingredients
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: allergen
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Ingredient list
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [Ingredients]
      summary: Create ingredient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientCreate"
      responses:
        "201":
          description: Ingredient created
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /ingredients/{id}:
    get:
      tags: [Ingredients]
      summary: Get ingredient by id
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Ingredient detail
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Ingredients]
      summary: Partially update ingredient
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientUpdate"
      responses:
        "200":
          description: Ingredient updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Ingredients]
      summary: Delete ingredient (admin role required)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Ingredient deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /recipes:
    get:
      tags: [Recipes]
      summary: List recipes
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: difficulty
          schema:
            type: string
            enum: [easy, medium, hard]
        - in: query
          name: createdBy
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Recipe list
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [Recipes]
      summary: Create recipe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeCreate"
      responses:
        "201":
          description: Recipe created
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /recipes/{id}:
    get:
      tags: [Recipes]
      summary: Get recipe by id
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Recipe detail
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Recipes]
      summary: Partially update recipe
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeUpdate"
      responses:
        "200":
          description: Recipe updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Recipes]
      summary: Delete recipe
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Recipe deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /analytics/recipes/{id}/nutrition:
    get:
      tags: [Analytics]
      summary: Nutrition analysis for a recipe
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Nutrition analysis
        "404":
          $ref: "#/components/responses/NotFound"
  /analytics/recipes/{id}/cost:
    get:
      tags: [Analytics]
      summary: Cost analysis for a recipe
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Cost analysis
        "404":
          $ref: "#/components/responses/NotFound"
  /analytics/leaderboard:
    get:
      tags: [Analytics]
      summary: Leaderboard by metric
      parameters:
        - in: query
          name: metric
          schema:
            type: string
            enum: [proteinDensity, costEfficiency]
            default: proteinDensity
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        "200":
          description: Ranked recipes
        "400":
          $ref: "#/components/responses/BadRequest"
  /analytics/recommendations:
    post:
      tags: [Analytics]
      summary: Recommend recipes by nutrition/cost/allergen constraints
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationRequest"
      responses:
        "200":
          description: Recommendation list
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Validation or domain constraint error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Authentication failed or missing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Role/permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource/route not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Duplicate or conflicting resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Request validation failed.
            details:
              nullable: true
      required: [success, error]
    RegisterRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 2
          maxLength: 80
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: Must include at least one letter and one number.
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 10
    IngredientCreate:
      type: object
      required: [name, category, kcalPer100g, proteinPer100g, carbsPer100g, fatPer100g, costPer100g]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        category:
          type: string
          minLength: 2
          maxLength: 60
        kcalPer100g:
          type: number
          minimum: 0
        proteinPer100g:
          type: number
          minimum: 0
        carbsPer100g:
          type: number
          minimum: 0
        fatPer100g:
          type: number
          minimum: 0
        costPer100g:
          type: number
          minimum: 0
        allergens:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 40
          default: []
        source:
          type: string
          minLength: 2
          maxLength: 120
          default: manual
    IngredientUpdate:
      type: object
      description: Partial update; at least one field is required by server validation.
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        category:
          type: string
          minLength: 2
          maxLength: 60
        kcalPer100g:
          type: number
          minimum: 0
        proteinPer100g:
          type: number
          minimum: 0
        carbsPer100g:
          type: number
          minimum: 0
        fatPer100g:
          type: number
          minimum: 0
        costPer100g:
          type: number
          minimum: 0
        allergens:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 40
        source:
          type: string
          minLength: 2
          maxLength: 120
    RecipeIngredient:
      type: object
      required: [ingredientId, grams]
      properties:
        ingredientId:
          type: integer
          minimum: 1
        grams:
          type: number
          exclusiveMinimum: true
          minimum: 0
          maximum: 5000
    RecipeCreate:
      type: object
      required: [title, description, difficulty, servings, prepMinutes, instructions, ingredients]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 120
        description:
          type: string
          minLength: 5
          maxLength: 600
        difficulty:
          type: string
          enum: [easy, medium, hard]
        servings:
          type: integer
          minimum: 1
          maximum: 50
        prepMinutes:
          type: integer
          minimum: 0
          maximum: 1440
        instructions:
          type: string
          minLength: 10
          maxLength: 4000
        ingredients:
          type: array
          minItems: 1
          maxItems: 30
          items:
            $ref: "#/components/schemas/RecipeIngredient"
    RecipeUpdate:
      type: object
      description: Partial update; at least one field is required by server validation.
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 120
        description:
          type: string
          minLength: 5
          maxLength: 600
        difficulty:
          type: string
          enum: [easy, medium, hard]
        servings:
          type: integer
          minimum: 1
          maximum: 50
        prepMinutes:
          type: integer
          minimum: 0
          maximum: 1440
        instructions:
          type: string
          minLength: 10
          maxLength: 4000
        ingredients:
          type: array
          minItems: 1
          maxItems: 30
          items:
            $ref: "#/components/schemas/RecipeIngredient"
    RecommendationRequest:
      type: object
      properties:
        targetCaloriesPerServing:
          type: number
          exclusiveMinimum: true
          minimum: 0
          maximum: 2000
        minProteinPerServing:
          type: number
          minimum: 0
          maximum: 300
        maxCostPerServing:
          type: number
          minimum: 0
          maximum: 100
        excludeAllergens:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 40
          default: []